Fixed bugs during parsing
- Dead event now checks for last death instead of single (can happen in fractals)
- Damage is now stored as long value
- Damage calculation filters out unknown npcs (could lead to a wrong parsing before)

+++ GW2RaidTool/EVTC-Log-Parser/Model/Data/Event/CombatEvent.cs
@@ -3,7 +3,7 @@
     public class CombatEvent
     {
 	    public int Time { get; set; }
-        public int Damage { get; set; }
+        public long Damage { get; set; }
         public int SkillId { get; set; }
         public bool IsBuff { get; set; }
         public Result Result { get; set; }
