+++ GW2RaidTool/EVTC-Log-Parser/Model/Data/Table/FinalDPS.cs
@@ -15,13 +15,13 @@
 		public string Profession { get; set; }
 		public double DPSAll => TotalAll / FightDuration;
 		public double DPSBoss => TotalBoss / FightDuration;
-		public int PowerAll { get; set; }
-		public int PowerBoss { get; set; }
-		public int CondiAll { get; set; }
-		public int CondiBoss { get; set; }
+		public long PowerAll { get; set; }
+		public long PowerBoss { get; set; }
+		public long CondiAll { get; set; }
+		public long CondiBoss { get; set; }
 
-		public int TotalAll => PowerAll + CondiAll;
-		public int TotalBoss => PowerBoss + CondiBoss;
+		public long TotalAll => PowerAll + CondiAll;
+		public long TotalBoss => PowerBoss + CondiBoss;
 
 		public int Down { get; set; }
 		public bool Dead { get; set; }
